Move map component under components folder
