+++ web-ng/src/app/components/main-page/main-page.module.ts
@@ -18,7 +18,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MainPageComponent } from './main-page.component';
 import { RouterModule } from '@angular/router';
-import { MapModule } from '../../../map/map.module';
+import { MapModule } from '../map/map.module';
 
 @NgModule({
   declarations: [MainPageComponent],
