+++ web-ng/src/app/components/map/map.module.ts
@@ -18,7 +18,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { MapComponent } from './map.component';
 import { AgmCoreModule } from '@agm/core';
-import { environment } from '../environments/environment';
+import { environment } from '../../../environments/environment';
 
 @NgModule({
   imports: [
